fix top-level `begin-for-syntax`
The interaction of the default eval handler and the compile handler is
a little different here than in current Racket. This defaval eval
handler not only detects and splits `begin`, but also
`begin-for-synatx`, calling the compile handler with the namespace set
to the one at the transformer phase, instead of passing a
`begin-for-syntax` form on to the compile handler. As a result,
`begin-for-syntax` content is compiled and evaluated in an interleaved
way, too.