add submodule manipulation
This change includes a revision to the way compiled code is encoded as
linklet directories, with an eye toward making submodules loadable
independently. A module is represented as a bundle, insteda of a
directory, and the intent is that directly layers will be flattened
into a lookup table, while bundles are loaded wholesale; a module
by itself is a bundle, while a module with its submodules is
a linklet directory.